% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_echo}
\alias{api_echo}
\title{#' @param data original data used for the query
#' @param query results form CTSgetR
#' @param from name of translated variable in the data
#' @param to name of data base translated to in the query
#' @export
#' @details combine query results with an existing data set
#' @import dplyr
CTSgetR_join<-function(data,query,from,to){
  
  #join
  by<-'searchTerm'
  names(by)<-from
  out<-left_join(data,
                 query,
                 by=by)
  
  #cleanup
  out<- out %>%
    select(one_of(c(colnames(data),'value')))
  colnames(out)[colnames(out) == 'value']<-make.names(to)
  
  return(out)
  
}}
\usage{
api_echo(...)
}
\description{
#' @param data original data used for the query
#' @param query results form CTSgetR
#' @param from name of translated variable in the data
#' @param to name of data base translated to in the query
#' @export
#' @details combine query results with an existing data set
#' @import dplyr
CTSgetR_join<-function(data,query,from,to){
  
  #join
  by<-'searchTerm'
  names(by)<-from
  out<-left_join(data,
                 query,
                 by=by)
  
  #cleanup
  out<- out %>%
    select(one_of(c(colnames(data),'value')))
  colnames(out)[colnames(out) == 'value']<-make.names(to)
  
  return(out)
  
}
}
