% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS.R
\name{CTS.options}
\alias{CTS.options}
\title{CTS.options}
\usage{

  CTS.options(url = "http://cts.fiehnlab.ucdavis.edu/service/conversion/toValues")
}
\arguments{
\item{url}{for options}
}
\details{
get translation options from CTS
}

