\name{CTSgetR-package}
\alias{CTSgetR-package}
\docType{package}
\title{
R interface to the Chemical Translation System (CTS)
}
\description{
Use the Chemical Translation System (CTS) to translate between > 250 of the most common biological database identifiers including: Chemical Name, InChIKey, InChI Code, PubChem CID, Pubchem SID, ChemDB, ChemSpider, Golm Metabolome Database (GMD), BioCyc, ChEBI, CAS, HMDB, KEGG, LipidMAPS.
}
\details{
\tabular{ll}{
Package: \tab CTSgetR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-06-03\cr
License: \tab TBA\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Dmitry Grapov
<dgrapov@ucdavis.edu>
}

%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{Chemical Translation System}
