% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS.R
\name{CTSgetR}
\alias{CTSgetR}
\title{CTSgetR}
\usage{
CTSgetR(id, from, to, limit.values = TRUE, progress = TRUE,
  server = "http://cts.fiehnlab.ucdavis.edu/service/convert")
}
\arguments{
\item{id}{a vector of metabolite identifier(s) or name(s) to translate see \code{\link{CTS.options}}}

\item{from}{Database name describing \code{id} see \code{\link{CTS.options}}}

\item{to}{Database name to translate \code{id} to see \code{\link{CTS.options}}
from    Database name describing \code{id} see \code{\link{CTS.options}}}

\item{limit.values}{return a single (first non empty) value or a comma separated string of multiple values (defaults to FALSE).}

\item{server}{Web address for the Chemical Translation System web services (defaults to http://cts.fiehnlab.ucdavis.edu/service/convert).}
}
\value{
data frame of results including \code{from} (fromIdentifier), \code{id} (searchTerm), \code{to} (toIdentifier) and translated values
}
\details{
Interface to CTS (http://cts.fiehnlab.ucdavis.edu/) for metabolite identifier translation between > 200 of the most common biological databases including: Chemical Name, InChIKey, PubChem CID, ChemSpider, BioCyc, ChEBI, CAS, HMDB, KEGG and LipidMAPS. Entries not found in CTS are returned as "", CTS errors are returned as 'error'. Observation:  CTS errors can be random, i.e. running the same query multiple times might result in success or error :(
}
\examples{
\dontrun{
id<-c("C15973","C00026","C05381")
from<-"KEGG" 
to<-"PubChem CID"
CTSgetR(id,from,to)
}
}
\seealso{
\code{\link{CTS.options}}
}

