% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_replace}
\alias{null_replace}
\title{#' @export
api_debug<-function(){
  
  out<-list(pwd=getwd(),
            db=dir('/ctsgetrdb'),
            date=Sys.time())
  
  
  jsonlite::toJSON(out)
  
}}
\usage{
null_replace(x, alt = NA)
}
\description{
#' @export
api_debug<-function(){
  
  out<-list(pwd=getwd(),
            db=dir('/ctsgetrdb'),
            date=Sys.time())
  
  
  jsonlite::toJSON(out)
  
}
}
