\name{Undocumented}

\alias{CIDtoName}
\alias{CTS.translate}
\alias{CTS.translate.async}
\alias{KEGGtoName}
\alias{NametoInchI}
\alias{NametoKEGG}
\alias{NametoKeyID}
\alias{NametoPubChem}
\alias{fixlc}
\alias{fixln}
\alias{getIDEOM}
\alias{test}


\title{
Undocumented functions
}
\description{
Collection of undocumented functions used internally.
}

\keyword{internal}