\name{CTSgetR-package}
\alias{CTSgetR-package}
\docType{package}
\title{
R interface to the Chemical Translation System (CTS)
}
\description{
Translate between > 200 of the most common metabolite database identifiers including: Chemical Name, InChIKey, InChI Code, PubChem CID, Pubchem SID, ChemDB, ChemSpider, Golm Metabolome Database (GMD), BioCyc, ChEBI, CAS, HMDB, KEGG, LipidMAPS.
}
\details{
\tabular{ll}{
Package: \tab CTSgetR\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2020-07-06\cr
License: \tab MIT\cr
}

}
\author{
Dmitry Grapov
<dgrapov@gmail.com>
}
\references{
http://cts.fiehnlab.ucdavis.edu/
}

